+++ navigator-compose/navigator-compose/src/androidTest/java/com/kpstv/navigation/compose/internels/Extensions.kt
@@ -1,0 +1,3 @@
+package com.kpstv.navigation.compose.internels
+
+internal fun<K, V> Map<K,V>.lastValue(): V? = get(keys.lastOrNull())